[chore] refactoring to more cleanly differentiate server and client code.
To resolve some issues with using bun (particularly since it is now in flux due to the rewrite), moves to using node as both build and deployment toolchain.  This necessitates the refactoring.
Adds missing translation string for select_task.domain.
