+++ frontend/Dockerfile
@@ -1,6 +1,5 @@
-FROM docker.io/oven/bun:slim AS base
+FROM docker.io/node:26-slim AS base
 RUN mkdir -p /app
-ENV NODE_ENV=production
 WORKDIR /app
 
 FROM base AS install
@@ -8,30 +7,28 @@ RUN mkdir -p /tmp/dev
 RUN mkdir -p /tmp/prod
 COPY ./package*.json /tmp/dev
 COPY ./package*.json /tmp/prod
-COPY ./bun.lock* /tmp/dev
-COPY ./bun.lock* /tmp/prod
 COPY ./patches /tmp/dev/patches
 COPY ./patches /tmp/prod/patches
-RUN cd /tmp/dev && bun install --frozen-lockfile
-RUN cd /tmp/prod && bun install --production --frozen-lockfile
+RUN cd /tmp/dev && npm ci
+RUN cd /tmp/prod && npm ci --omit=dev
 
-FROM base AS prerelease
+FROM base AS build
 COPY --from=install /tmp/dev/node_modules node_modules
 COPY . .
-RUN bun run build
-RUN bun run build:server
+ENV NODE_ENV=production
+RUN npm run build
+RUN npm run build:server
 
-FROM docker.io/node:26-slim AS release
-# Using node image as bun image has had issues with native modules in some cases
-# TODO: check if bun can be used directly here with ltijs security module and other native modules
+FROM base AS release
 LABEL org.opencontainers.image.description="myProse web application"
 LABEL org.opencontainers.image.licenses=CC-BY-NC-SA-4.0
 EXPOSE 8888
 ENV PORT=8888
+ENV NODE_ENV=production
 WORKDIR /app
 COPY --from=install /tmp/prod/node_modules node_modules
-COPY --from=prerelease /app/dist/ ./dist/
-COPY --from=prerelease /app/build/ ./build/
-COPY --from=prerelease /app/public/ ./public/
-COPY --from=prerelease /app/package.json ./package.json
+COPY --from=build /app/dist/ ./dist/
+COPY --from=build /app/build/ ./build/
+COPY --from=build /app/public/ ./public/
+COPY --from=build /app/package.json ./package.json
 CMD [ "node", "build/server/index.js" ]
