+++ frontend/package.json
@@ -14,12 +14,12 @@
   },
   "private": true,
   "scripts": {
-    "dev": "npx bun ./index.ts",
-    "build": "npx -y bunx --bun vike build",
-    "build:server": "npx bun build index.ts --outdir=./build/server --target=node",
+    "dev": "npx tsx ./index.ts",
+    "build": "vike build",
+    "build:server": "vite build --config vite.config.server.ts",
     "test": "npx bun test",
     "test:watch": "npx bun test --watch",
-    "preview": "cross-env NODE_ENV=production npx bun ./index.ts",
+    "preview": "cross-env NODE_ENV=production npm run build && npx tsx ./index.ts",
     "start": "npm run preview",
     "check_format": "prettier -c src/ pages/ components/ layouts/ index.ts",
     "lint": "eslint src pages components layouts index.ts",
@@ -29,7 +29,8 @@
     "gen:schema": "npx typescript-json-schema ./tsconfig.gen.json WritingTask --out ./src/lib/WritingTask.schema.json",
     "storybook": "storybook dev -p 6006",
     "upgrade-storybook": "npx storybook@latest upgrade",
-    "build-storybook": "storybook build"
+    "build-storybook": "storybook build",
+    "postinstall": "patch-package"
   },
   "repository": {
     "type": "git",
@@ -58,17 +59,17 @@
   ],
   "devDependencies": {
     "@chromatic-com/storybook": "^5.2.1",
-    "@eslint-react/eslint-plugin": "^5.14.10",
+    "@eslint-react/eslint-plugin": "^5.17.3",
     "@eslint/js": "^10.0.1",
-    "@happy-dom/global-registrator": "^20.10.6",
+    "@happy-dom/global-registrator": "^20.11.0",
     "@rollup/plugin-commonjs": "^29.0.3",
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^16.0.3",
     "@rollup/plugin-typescript": "^12.3.0",
-    "@storybook/addon-docs": "^10.5.0",
-    "@storybook/addon-onboarding": "^10.5.0",
-    "@storybook/addon-vitest": "^10.5.0",
-    "@storybook/react-vite": "^10.5.0",
+    "@storybook/addon-docs": "^10.5.3",
+    "@storybook/addon-onboarding": "^10.5.3",
+    "@storybook/addon-vitest": "^10.5.3",
+    "@storybook/react-vite": "^10.5.3",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
@@ -91,37 +92,40 @@
     "@types/react-dom": "^19.2.3",
     "@types/sanitize-html": "^2.16.1",
     "@types/string-format": "^2.0.3",
+    "@typescript/native": "npm:typescript@^7.0.2",
     "@vitejs/plugin-react": "^6.0.3",
     "@vitest/browser-playwright": "^4.1.10",
     "@vitest/coverage-v8": "^4.1.10",
-    "autoprefixer": "^10.5.3",
+    "autoprefixer": "^10.5.4",
     "bun-plugin-svgr": "^0.1.0",
     "cross-env": "^10.1.0",
-    "dotenv": "^17.4.2",
     "eslint": "^10.7.0",
-    "eslint-plugin-storybook": "^10.5.0",
+    "eslint-plugin-storybook": "^10.5.3",
     "globals": "^17.7.0",
     "playwright": "^1.61.1",
-    "postcss": "^8.5.19",
+    "postcss": "^8.5.20",
     "prettier": "^3.9.5",
     "rollup-plugin-visualizer": "^7.0.1",
     "sass": "^1.101.0",
-    "storybook": "^10.5.0",
+    "standaloner": "^0.3.3",
+    "storybook": "^10.5.3",
+    "typescript": "npm:@typescript/typescript6@^6.0.2",
     "typescript-eslint": "^8.64.0",
-    "vite": "^8.1.4",
+    "vite": "^8.1.5",
     "vite-plugin-svgr": "^5.2.0",
     "vitest": "^4.1.10",
     "vitest-fetch-mock": "^0.4.5"
   },
   "dependencies": {
-    "@anthropic-ai/sdk": "^0.111.0",
-    "@fortawesome/fontawesome-free": "^7.3.0",
-    "@fortawesome/fontawesome-svg-core": "^7.3.0",
-    "@fortawesome/free-regular-svg-icons": "^7.3.0",
-    "@fortawesome/free-solid-svg-icons": "^7.3.0",
-    "@fortawesome/react-fontawesome": "^3.4.0",
+    "@anthropic-ai/sdk": "^0.112.3",
+    "@fortawesome/fontawesome-free": "^7.3.1",
+    "@fortawesome/fontawesome-svg-core": "^7.3.1",
+    "@fortawesome/free-regular-svg-icons": "^7.3.1",
+    "@fortawesome/free-solid-svg-icons": "^7.3.1",
+    "@fortawesome/react-fontawesome": "^3.5.0",
     "@googleworkspace/drive-picker-element": "^0.7.3",
-    "@tanstack/react-query": "^5.101.2",
+    "@tanstack/react-query": "^5.101.3",
+    "@vikejs/express": "^0.2.4",
     "ajv": "^8.20.0",
     "better-auth": "^1.6.23",
     "bootstrap": "^5.3.8",
@@ -136,6 +140,7 @@
     "datatables.net-react": "^1.0.2",
     "docx": "^9.7.1",
     "dompurify": "^3.4.12",
+    "dotenv": "^17.4.2",
     "escape-html": "^1.0.3",
     "express": "^5.2.1",
     "express-basic-auth": "^1.2.1",
@@ -152,13 +157,14 @@
     "mongodb": "^7.5.0",
     "multer": "^2.2.0",
     "node-html-parser": "^9.0.0",
+    "patch-package": "^8.0.1",
     "prom-client": "^15.1.3",
     "react": "^19.2.7",
     "react-animate-height": "^3.2.4",
     "react-bootstrap": "^2.10.10",
     "react-dom": "^19.2.7",
     "react-error-boundary": "^6.1.2",
-    "react-i18next": "^17.0.9",
+    "react-i18next": "^17.0.10",
     "react-split": "^2.0.14",
     "react-streaming": "^0.4.20",
     "sanitize-html": "^2.17.6",
@@ -188,12 +194,8 @@
     "minimatch": ">=10.2.1",
     "lodash": ">4.17.22"
   },
-  "patchedDependencies": {
-    "ltijs@5.9.9": "patches/ltijs@5.9.9.patch",
-    "bson@7.3.1": "patches/bson@7.3.1.patch"
-  },
   "allowScripts": {
-    "esbuild@0.28.1": true,
-    "@parcel/watcher@2.5.6": true
+    "@parcel/watcher": true,
+    "esbuild": true
   }
 }
