+++ frontend/patches/ltijs+5.9.9.patch
@@ -1,7 +1,7 @@
-diff --git a/dist/Provider/Services/DeepLinking.js b/dist/Provider/Services/DeepLinking.js
-index a665b036f4a0a3d139356edd0684c7ad8d2cc5dd..08b7aead58454f1a86389b740422aa71be5039d4 100644
---- a/dist/Provider/Services/DeepLinking.js
-+++ b/dist/Provider/Services/DeepLinking.js
+diff --git a/node_modules/ltijs/dist/Provider/Services/DeepLinking.js b/node_modules/ltijs/dist/Provider/Services/DeepLinking.js
+index a665b03..b6455e4 100644
+--- a/node_modules/ltijs/dist/Provider/Services/DeepLinking.js
++++ b/node_modules/ltijs/dist/Provider/Services/DeepLinking.js
 @@ -9,11 +9,11 @@ function _assertClassBrand(e, t, n) { if ("function" == typeof e ? e === t : e.h
  
  const jwt = require('jsonwebtoken');
@@ -13,7 +13,7 @@ index a665b036f4a0a3d139356edd0684c7ad8d2cc5dd..08b7aead58454f1a86389b740422aa71
 -const deepLinkSubmissionForm = path.join(__dirname, '../../Templates', 'DeepLinkSubmissionForm.html');
 +// const path = require('path');
 +// const {
-+//   sprightly
++  // sprightly
 +// } = require('sprightly');
 +// const deepLinkSubmissionForm = path.join(__dirname, '../../Templates', 'DeepLinkSubmissionForm.html');
  var _getPlatform = /*#__PURE__*/new WeakMap();
@@ -28,8 +28,8 @@ index a665b036f4a0a3d139356edd0684c7ad8d2cc5dd..08b7aead58454f1a86389b740422aa71
 -      message
 -    });
 +    // const form = sprightly(deepLinkSubmissionForm, {
-+    //   action: idtoken.platformContext.deepLinkingSettings.deep_link_return_url,
-+    //   message
++      // action: idtoken.platformContext.deepLinkingSettings.deep_link_return_url,
++      // message
 +    // });
 +    const form = `<form id="ltijs_submit" style="display: none;" action="${idtoken.platformContext.deepLinkingSettings.deep_link_return_url}" method="POST">
 +  <input type="hidden" name="JWT" value="${message}" />
@@ -40,10 +40,10 @@ index a665b036f4a0a3d139356edd0684c7ad8d2cc5dd..08b7aead58454f1a86389b740422aa71
      return form;
    }
  
-diff --git a/dist/Utils/Database.js b/dist/Utils/Database.js
-index d48737e905fdcc8d6dac7650478520c4a4bd3e54..c9eb5174d083c068efc8265d1283415528c4fed2 100644
---- a/dist/Utils/Database.js
-+++ b/dist/Utils/Database.js
+diff --git a/node_modules/ltijs/dist/Utils/Database.js b/node_modules/ltijs/dist/Utils/Database.js
+index d48737e..c9eb517 100644
+--- a/node_modules/ltijs/dist/Utils/Database.js
++++ b/node_modules/ltijs/dist/Utils/Database.js
 @@ -24,9 +24,9 @@ class Database {
    constructor(database) {
      _classPrivateFieldInitSpec(this, _dbUrl, void 0);
