+++ frontend/src/client/slate.ts
@@ -11,7 +11,7 @@ import {
 import escapeHtml from 'escape-html';
 import { type Descendant, Node as SlateNode, Text } from 'slate';
 import { jsx } from 'slate-hyperscript';
-import type { WritingTask } from './WritingTask';
+import type { WritingTask } from '../lib/WritingTask';
 import type { CustomElement, CustomText, ListElement } from '../slate';
 
 /**
