+++ frontend/tsconfig.server.json
@@ -3,12 +3,21 @@
   "compilerOptions": {
     "esModuleInterop": true,
     "lib": ["ESNext"],
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "resolveJsonModule": true
+    "module": "esnext",
+    "moduleResolution": "nodenext",
+    "outDir": "./build/server",
+    "paths": {
+      "#/*": ["./src/*"],
+      "#assets/*": ["./assets/*"],
+      "#lib/*": ["./src/lib/*"],
+      "#server/*": ["./src/server/*"]
+    },
+    "resolveJsonModule": true,
+    "target": "es2025"
   },
   "ts-node": {
     "esm": true
   },
   "include": ["./src/server/**.ts", "./src/lib/**.ts", "./vike.d.ts", "./src/ltijs.d.ts", "./src/index.d.ts", "./index.ts"],
+  "exclude": ["node_modules", "build", "dist", "./src/**.test.*", "./src/**.spec.*"]
 }
