+++ frontend/vite.config.server.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from 'vite';
+import { builtinModules } from 'module';
+
+export default defineConfig(({ mode }) => ({
+  build: {
+    ssr: './index.ts',
+    outDir: './build/server',
+  },
+  ssr: {
+    external: [
+      ...builtinModules,
+      ...builtinModules.map((m) => `node:${m}`),
+    ]
+  }
+}));
